Refactoring
Made new packages and refactored classes
