+++ src/main/java/org/wecancodeit/reviews/Populator.java
@@ -2,8 +2,10 @@ package org.wecancodeit.reviews;
 
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
-import org.wecancodeit.reviews.models.CategoryStorage;
-import org.wecancodeit.reviews.models.ReviewStorage;
+import org.wecancodeit.reviews.models.Category;
+import org.wecancodeit.reviews.models.Review;
+import org.wecancodeit.reviews.storage.CategoryStorage;
+import org.wecancodeit.reviews.storage.ReviewStorage;
 
 @Component
 public class Populator implements CommandLineRunner {
