+++ src/main/java/org/wecancodeit/reviews/controllers/CategoriesController.java
@@ -3,7 +3,7 @@ package org.wecancodeit.reviews.controllers;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.wecancodeit.reviews.models.CategoryStorage;
+import org.wecancodeit.reviews.storage.CategoryStorage;
 
 @Controller
 @RequestMapping("categories")
