+++ src/main/java/org/wecancodeit/reviews/controllers/ReviewController.java
@@ -5,8 +5,8 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.wecancodeit.reviews.Review;
-import org.wecancodeit.reviews.models.ReviewStorage;
+import org.wecancodeit.reviews.models.Review;
+import org.wecancodeit.reviews.storage.ReviewStorage;
 
 
 @Controller
