+++ src/main/java/org/wecancodeit/reviews/storage/HashtagStorageJpaImpl.java
@@ -1,8 +1,8 @@
-package org.wecancodeit.reviews;
+package org.wecancodeit.reviews.storage;
 
 import org.springframework.stereotype.Service;
-import org.wecancodeit.reviews.models.HashtagRepository;
-import org.wecancodeit.reviews.models.HashtagStorage;
+import org.wecancodeit.reviews.models.Hashtag;
+import org.wecancodeit.reviews.storage.repositories.HashtagRepository;
 
 import java.util.Collection;
