+++ src/main/java/org/wecancodeit/reviews/storage/ReviewStorage.java
@@ -1,6 +1,6 @@
-package org.wecancodeit.reviews.models;
+package org.wecancodeit.reviews.storage;
 
-import org.wecancodeit.reviews.Review;
+import org.wecancodeit.reviews.models.Review;
 
 import java.util.Collection;
