+++ src/main/java/org/wecancodeit/reviews/storage/repositories/HashtagRepository.java
@@ -1,7 +1,7 @@
-package org.wecancodeit.reviews.models;
+package org.wecancodeit.reviews.storage.repositories;
 
 import org.springframework.data.repository.CrudRepository;
-import org.wecancodeit.reviews.Hashtag;
+import org.wecancodeit.reviews.models.Hashtag;
 
 public interface HashtagRepository extends CrudRepository <Hashtag,Long>
